voice_agent/requirements.txt
# ARC Voice Agent — Python dependencies
# Install: pip install -r requirements.txt

# LiveKit Agents framework + Google (Gemini) plugin
livekit-agents[google,silero]>=0.12.0

# LiveKit API (for token generation)
livekit-api>=0.8.0

# Token server
fastapi>=0.110.0
uvicorn[standard]>=0.29.0

# Env loading
python-dotenv>=1.0.0
